move previous autogen integration to ag2
