+++ docs/docs/tracing/index.mdx
@@ -187,6 +187,9 @@ MLflow Tracing is integrated with various GenAI libraries and provide **one-line
   <SmallLogoCard link="/tracing/integrations/autogen">
     <span>![AutoGen Logo](/images/logos/autogen-logo.png)</span>
   </SmallLogoCard>
+  <SmallLogoCard link="/tracing/integrations/ag2">
+    <span>![AG2 Logo](/images/logos/ag2-logo.png)</span>
+  </SmallLogoCard>
   <SmallLogoCard link="/tracing/integrations/gemini">
     <span>![Gemini Logo](/images/logos/google-gemini-logo.svg)</span>
   </SmallLogoCard>
