+++ docs/docs/tracing/integrations/ag2.mdx
@@ -1,6 +1,6 @@
 ---
 sidebar_position: 6
-sidebar_label: AutoGen
+sidebar_label: AG2
 ---
 
 import { APILink } from "@site/src/components/APILink";
@@ -9,20 +9,21 @@ import TOCInline from "@theme/TOCInline";
 import Tabs from "@theme/Tabs";
 import TabItem from "@theme/TabItem";
 
-# Tracing AutoGen🤖
+# Tracing AG2🤖
 
-{/* TODO: rewrite this page for AutoGen 0.4 */}
-{/* ![AutoGen Tracing via autolog](/images/llms/autogen/autogen-trace.png) */}
+![AutoGen Tracing via autolog](/images/llms/ag2/ag2-trace.png)
 
-[AutoGen](https://microsoft.github.io/autogen/0.2/) is an open-source framework for building event-driven, distributed, scalable, and resilient AI agent systems.
+[AG2](https://ag2.ai/) is an open-source framework for building and orchestrating AI agent interactions.
 
-[MLflow Tracing](../) provides automatic tracing capability for AutoGen, an open-source multi-agent framework. By enabling auto tracing
-for AutoGen by calling the <APILink fn="mlflow.autogen.autolog" /> function, MLflow will capture nested traces and log them to the active MLflow Experiment upon agents execution.
+[MLflow Tracing](../) provides automatic tracing capability for AG2, an open-source multi-agent framework. By enabling auto tracing
+for AG2 by calling the <APILink fn="mlflow.ag2.autolog" /> function, MLflow will capture nested traces and logged them to the active MLflow Experiment upon agents execution.
+Note that since AG2 is built based on [AutoGen 0.2](https://microsoft.github.io/autogen/0.2/), this integration can be used when you use AutoGen 0.2.
+
 
 ```python
 import mlflow
 
-mlflow.autogen.autolog()
+mlflow.ag2.autolog()
 ```
 
 MLflow captures the following information about the multi-agent execution:
@@ -43,15 +44,15 @@ from autogen import ConversableAgent
 
 import mlflow
 
-# Turn on auto tracing for AutoGen
-mlflow.autogen.autolog()
+# Turn on auto tracing for AG2
+mlflow.ag2.autolog()
 
 # Optional: Set a tracking URI and an experiment
 mlflow.set_tracking_uri("http://localhost:5000")
-mlflow.set_experiment("AutoGen")
+mlflow.set_experiment("AG2")
 
 
-# Define a simple multi-agent workflow using AutoGen
+# Define a simple multi-agent workflow using AG2
 config_list = [
     {
         "model": "gpt-4o-mini",
@@ -107,4 +108,4 @@ response = user_proxy.initiate_chat(
 
 ### Disable auto-tracing
 
-Auto tracing for AutoGen can be disabled globally by calling `mlflow.autogen.autolog(disable=True)` or `mlflow.autolog(disable=True)`.
+Auto tracing for AG2 can be disabled globally by calling `mlflow.ag2.autolog(disable=True)` or `mlflow.autolog(disable=True)`.
