+++ mlflow/utils/autologging_utils/__init__.py
@@ -72,6 +72,7 @@ _AUTOLOGGING_SUPPORTED_VERSION_WARNING_SUPPRESS_LIST = [
     "openai",
     "dspy",
     "autogen",
+    "ag2",
     "gemini",
     "anthropic",
     "crewai",
